requirements.txt
Django==1.4
South==0.7.6
dj-database-url==0.2.1
django-mailer==0.1.0
django-social-auth==0.7.0
django-storages==1.1.4
gunicorn==0.16.1
httplib2==0.7.7
oauth2==1.5.211
psycopg2==2.4.1
python-openid==2.2.5
selenium==2.42.1
wsgiref==0.1.2
boto
